Move ISelection out of editorCommon
