+++ src/vs/workbench/api/node/extHostTypeConverters.ts
@@ -8,12 +8,13 @@ import Severity from 'vs/base/common/severity';
 import * as modes from 'vs/editor/common/modes';
 import * as types from './extHostTypes';
 import { Position as EditorPosition } from 'vs/platform/editor/common/editor';
-import { ISelection, IDecorationOptions, EndOfLineSequence } from 'vs/editor/common/editorCommon';
+import { IDecorationOptions, EndOfLineSequence } from 'vs/editor/common/editorCommon';
 import * as vscode from 'vscode';
 import URI from 'vs/base/common/uri';
 import { SaveReason } from 'vs/workbench/services/textfile/common/textfiles';
 import { IPosition } from "vs/editor/common/core/position";
 import { IRange } from "vs/editor/common/core/range";
+import { ISelection } from "vs/editor/common/core/selection";
 
 export interface PositionLike {
 	line: number;
